+++ Jmol/src/org/openscience/jmol/VectorShape.java
@@ -45,8 +45,8 @@ class VectorShape implements Shape, Transformable {
 
   }
 
-  public double getZ() {
-    return screenPositionEnd.z + 4.0;
+  public int getZ() {
+    return (int)screenPositionEnd.z + 4;
   }
 
   /**
